requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
werkzeug==3.0.4; python_version >= '3.8'
